[STYLE] NetworkError departs from API, unnecessary keyword ObservableObject removed
